refactor: 설정 파일을 config 패키지로 이동
