+++ src/main/java/wooteco/subway/config/WebConfig.java
@@ -1,4 +1,4 @@
-package wooteco.subway;
+package wooteco.subway.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
